// front/src/utils/types.ts
interface PreviewWidth {
	header: number;
	headerMin520: number;
	headerMin600: number;
	headerMin710: number;
	headerMin900: number;
	headerMin1200: number;
	headerGrid: number;
	content: number;
}

interface PreviewMarginLeft {
	header: number;
	headerMin520: number;
	headerMin600: number;
	headerMin710: number;
	headerMin900: number;
	headerMin1200: number;
}

interface PreviewDividerWidth {
	header: string;
	headerMin520: string;
	headerMin600: string;
	headerMin710: string;
	headerMin900: string;
	headerMin1200: string;
}

export type { PreviewWidth, PreviewMarginLeft, PreviewDividerWidth };